pyproject.toml
[project]
name = "11px"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.6.0",
    "fastapi>=0.115.14",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
